+++ src/SwooleRequestHandlerRunner.php
@@ -15,16 +15,17 @@ use Psr\Http\Server\RequestHandlerInterface;
 use Swoole\Http\Request as SwooleHttpRequest;
 use Swoole\Http\Response as SwooleHttpResponse;
 use Swoole\Http\Server as SwooleHttpServer;
-use Swoole\Process as SwooleProcess;
 use Throwable;
 use Zend\Expressive\Swoole\Exception;
 use Zend\HttpHandlerRunner\Emitter\EmitterInterface;
 use Zend\HttpHandlerRunner\RequestHandlerRunner;
 
-use function date;
-use function microtime;
-use function time;
-use function usleep;
+use function chdir;
+use function getcwd;
+use function sprintf;
+use function swoole_set_process_name;
+
+use const PHP_OS;
 
 /**
  * "Run" a request handler using Swoole.
@@ -39,6 +40,11 @@ use function usleep;
  */
 class SwooleRequestHandlerRunner extends RequestHandlerRunner
 {
+    /**
+     * Default Process Name
+     */
+    public const DEFAULT_PROCESS_NAME = 'expressive';
+
     /**
      * Keep CWD in daemon mode.
      *
@@ -97,6 +103,11 @@ class SwooleRequestHandlerRunner extends RequestHandlerRunner
      */
     private $staticResourceHandler;
 
+    /**
+     * @var string
+     */
+    private $processName;
+
     public function __construct(
         RequestHandlerInterface $handler,
         callable $serverRequestFactory,
@@ -104,7 +115,8 @@ class SwooleRequestHandlerRunner extends RequestHandlerRunner
         PidManager $pidManager,
         SwooleHttpServer $httpServer,
         StaticResourceHandlerInterface $staticResourceHandler = null,
-        Log\AccessLogInterface $logger = null
+        Log\AccessLogInterface $logger = null,
+        string $processName = self::DEFAULT_PROCESS_NAME
     ) {
         $this->handler = $handler;
 
@@ -129,6 +141,7 @@ class SwooleRequestHandlerRunner extends RequestHandlerRunner
             new Log\StdoutLogger(),
             new Log\AccessLogFormatter()
         );
+        $this->processName = $processName;
         $this->cwd = getcwd();
     }
 
@@ -160,6 +173,7 @@ class SwooleRequestHandlerRunner extends RequestHandlerRunner
 
         // Reset CWD
         chdir($this->cwd);
+        $this->setProcessName(sprintf('%s-master', $this->processName));
 
         $this->logger->notice('Swoole is running at {host}:{port}, in {cwd}', [
             'host' => $server->host,
@@ -177,6 +191,11 @@ class SwooleRequestHandlerRunner extends RequestHandlerRunner
     {
         // Reset CWD
         chdir($this->cwd);
+
+        $processName = $workerId >= $server->setting['worker_num']
+            ? sprintf('%s-task-worker-%d', $this->processName, $workerId)
+            : sprintf('%s-worker-%d', $this->processName, $workerId);
+        $this->setProcessName($processName);
 
         $this->logger->notice('Worker started in {cwd} with ID {pid}', [
             'cwd' => getcwd(),
@@ -215,7 +234,7 @@ class SwooleRequestHandlerRunner extends RequestHandlerRunner
         $this->logger->logAccessForPsr7Resource($request, $psr7Response);
     }
 
-     /**
+    /**
      * Handle the shutting down of the server
      */
     public function onShutdown(SwooleHttpServer $server) : void
@@ -235,5 +254,18 @@ class SwooleRequestHandlerRunner extends RequestHandlerRunner
         $psr7Response = ($this->serverRequestErrorResponseGenerator)($exception);
         $emitter->emit($psr7Response);
         $this->logger->logAccessForPsr7Resource($request, $psr7Response);
+    }
+
+    /**
+     * Set the process name, only if the current OS supports the operation
+     *
+     * @param string $name
+     */
+    private function setProcessName(string $name) : void
+    {
+        if (PHP_OS === 'Darwin') {
+            return;
+        }
+        swoole_set_process_name($name);
     }
 }